Backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic[email]
passlib[bcrypt]
python-multipart
python-dotenv
python-jose[cryptography]
psycopg2-binary
